+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Added method ConfigRuntimeException.convertToRuntimeEx()
+- Added CloseableDAOContextAbstract.newCloseableDAOContextDS()
+- Added CloseableDAOContextAbstract.newCloseableDAOContextCF()
+
 ## [8.6.5] - 2024-09-07
 
 ### Added